agent/pyproject.toml
[project]
name = "agentic-rag"
version = "0.1.0"
description = "Agentic RAG Demo - Document management and retrieval using Strands Agents"
requires-python = ">=3.11"
dependencies = [
    "strands-agents[bedrock]>=0.1.0",
    "bedrock-agentcore>=0.1.0",
    "boto3>=1.38.0",
    "pypdf>=5.0.0",
    "python-docx>=1.0.0",
    "python-dotenv>=1.0.0",
    "streamlit>=1.30.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
    "mypy>=1.10.0",
    "ruff>=0.5.0",
    "moto[s3]>=5.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/agentic_rag"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "SIM"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
